Реализовал класс-родитель Component
